添加 jpa demo 实例，添加redis缓存支持,修改springboot包版本以增加 自动配置cache

+++ src/main/java/com/xwb/demo/repository/PeopleRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.xwb.demo.meta.People;
 
-public interface PeopleRepository extends JpaRepository<People, Integer>{
+public interface PeopleRepository extends JpaRepository<People, Long>{
 
 }
